Refactored types into a different file
